+++ src/lib/core.js
@@ -249,6 +249,7 @@ var balloon = {
       label: 'menu.trash',
       icon: 'trash',
       callback: function() {
+        balloon.tree.filter.deleted = 2;
         return balloon.refreshTree('/nodes/trash', {}, {});
       }
     },
@@ -2296,6 +2297,11 @@ var balloon = {
    * @return  void
    */
   _userProfileNavigateTo: function(view) {
+    var $curActive = $('#fs-profile-window dl dt.active');
+
+    // Do not navigate to currently active
+    if($curActive.length > 0 && $curActive.attr('id').substr(24) === view) return;
+
     $('#fs-profile-window dl > *').removeClass('active');
 
     $('#fs-profile-window-title-'+view).addClass('active');
@@ -2368,6 +2374,9 @@ var balloon = {
 
         // User attributes
         var $table = $('#fs-profile-user').find('table');
+        // make sure no double data is displayed #244
+        $table.find('tr').remove();
+
         var attributes = ['id', 'username', 'created'];
         for(var i=0; i<attributes.length; i++) {
           var attribute = attributes[i];
@@ -3618,7 +3627,7 @@ var balloon = {
             ts = Math.round(date.getTime() / 1000);
           }
 
-          balloon._setDestroy(node, ts, $k_win);
+          balloon._setDestroy(node, ts, $k_win, true);
         });
 
         $fs_destroy_date_win.find('input:button').unbind().click(function() {
@@ -3636,17 +3645,17 @@ var balloon = {
    * @param object node
    * @param null|integer ts
    */
-  _setDestroy: function(node, ts, $k_win) {
+  _setDestroy: function(node, ts, $k_win, reload) {
     var curTs = (new Date(node.destroy)).getTime() / 1000;
     var $d = $.Deferred();
 
     if(ts !== curTs) {
       if(ts === null) {
-        balloon.selfDestroyNode(node, ts, $k_win, $d);
+        balloon.selfDestroyNode(node, ts, $k_win, $d, reload);
       } else {
         var dateHr = kendo.toString(new Date(ts * 1000), kendo.culture().calendar.patterns.g)
         var msg  = i18next.t('view.properties.prompt_destroy', dateHr, node.name);
-        balloon.promptConfirm(msg, 'selfDestroyNode', [node, ts, $k_win]).then(function() {
+        balloon.promptConfirm(msg, 'selfDestroyNode', [node, ts, $k_win, reload]).then(function() {
           $d.resolve();
         }, function() {
           $d.resolve();
@@ -3669,7 +3678,7 @@ var balloon = {
    * @param object node
    * @param null|integer ts
    */
-  selfDestroyNode: function(node, ts, $k_win) {
+  selfDestroyNode: function(node, ts, $k_win, reload) {
     var url;
 
     var $d = $.Deferred();
@@ -3688,10 +3697,18 @@ var balloon = {
           $k_win.close();
         }
 
-        $d.resolve();
+        if(reload) {
+          balloon.reloadTree().then(function() {
+            if(balloon.last === null) return;
 
-        node.destroy = ts === null ? undefined : (new Date(ts * 1000)).toISOString();
-        balloon.displayProperties(node);
+            balloon._updateContentView(balloon.last);
+            balloon.switchView('properties');
+
+            $d.resolve();
+          });
+        } else {
+          $d.resolve();
+        }
       }
     });
 
@@ -6031,14 +6048,12 @@ var balloon = {
       dataType: 'json',
       statusCode: {
         204: function(e) {
-          balloon.refreshTree('/collections/children', {id: balloon.getCurrentCollectionId()});
+          balloon.reloadTree().then(function() {
+            if(balloon.last === null || node.shareowner.id !== login.user.id) return;
 
-          if(node.shareowner.id === login.user.id) {
-            balloon.last.shared = false;
-            if(balloon.id(node) == balloon.id(balloon.last)) {
-              balloon.switchView('share');
-            }
-          }
+            balloon._updateContentView(balloon.last);
+            balloon.switchView('share');
+          });
         }
       },
     });
@@ -6065,12 +6080,12 @@ var balloon = {
         name: name
       },
       success: function(data) {
-        node.shared = true;
-        balloon.reloadTree();
-        if(balloon.id(node) == balloon.id(balloon.last)) {
-          balloon.last.shareowner = data.shareowner;
+        balloon.reloadTree().then(function() {
+          if(balloon.last === null) return;
+
+          balloon._updateContentView(balloon.last);
           balloon.switchView('share');
-        }
+        });
       },
     };
 
@@ -6295,9 +6310,14 @@ var balloon = {
 
           $.when(shareLinkRequest, destroyRequest).done(function() {
             balloon.reloadTree().then(function() {
-              balloon.switchView('share-link');
+              if(balloon.last === null) return;
+
               $k_win.close();
               balloon.showShareLink();
+
+              balloon._updateContentView(balloon.last);
+              balloon.switchView('share-link');
+
             });
           });
         });
@@ -6396,12 +6416,16 @@ var balloon = {
       url: balloon.base+'/nodes/share-link?id='+balloon.id(node),
       type: 'DELETE',
       success: function(body) {
-        delete balloon.last.sharelink_token;
-        balloon.reloadTree();
+        balloon.reloadTree().then(function() {
+          if(balloon.last === null) return;
 
-        $('#fs-share-link-edit').hide();
-        $('#fs-share-link-delete').hide();
-        $('#fs-share-link-create').show();
+          $('#fs-share-link-edit').hide();
+          $('#fs-share-link-delete').hide();
+          $('#fs-share-link-create').show();
+
+          balloon._updateContentView(balloon.last);
+          balloon.switchView('share-link');
+        });
       }
     });
   },